+++ server/src/compiler/utilities.ts
@@ -204,8 +204,6 @@ export function createDetachedDiagnostic(fileName: string, diagSourceText: strin
         text = formatStringFromArgs(text, args);
     }
 
-    if (start == 2512 && fileName.endsWith("hall02.c")) debugger;
-
     return {
         file: undefined,
         start,